+++ lib/blind_index.rb
@@ -8,53 +8,62 @@ require "blind_index/version"
 module BlindIndex
   class Error < StandardError; end
 
-  def self.generate_bidx(value, key:, iterations: 10000, expression: nil, algorithm: :pbkdf2_hmac, **options)
-    key = key.call if key.respond_to?(:call)
+  class << self
+    attr_accessor :default_options
+  end
+  self.default_options = {
+    iterations: 10000,
+    algorithm: :pbkdf2_hmac,
+    insecure_key: false,
+    encode: true
+  }
 
-    raise BlindIndex::Error, "Missing key for blind index" unless key
-    key = key.to_str
+  def self.generate_bidx(value, key:, **options)
+    options = default_options.merge(options)
 
     # apply expression
-    value = expression.call(value) if expression
+    value = options[:expression].call(value) if options[:expression]
 
     unless value.nil?
+      algorithm = options[:algorithm].to_sym
+
+      key = key.call if key.respond_to?(:call)
+      raise BlindIndex::Error, "Missing key for blind index" unless key
+
+      key = key.to_s
+      unless options[:insecure_key] && algorithm == :pbkdf2_hmac
+        raise BlindIndex::Error, "Key must use binary encoding" if key.encoding != Encoding::BINARY
+        # raise BlindIndex::Error, "Key must not be ASCII" if key.bytes.all? { |b| b < 128 }
+        raise BlindIndex::Error, "Key must be 32 bytes" if key.bytesize != 32
+      end
+
+      # gist to compare algorithm results
+      # https://gist.github.com/ankane/fe3ac63fbf1c4550ee12554c664d2b8c
       value =
-        case algorithm.to_sym
+        case algorithm
         when :scrypt
-          validate_key(key, 32)
-
-          # libsodium equivalent
-          # use this script to convert values
-          # https://gist.github.com/ankane/5008469d8b25fb614e84883c9285994a
-          # RbNaCl::PasswordHash.scrypt(value.to_s, key, 2**20, 2**24, 32)
-
+          # n, p (keep r at 8)
           SCrypt::Engine.scrypt(value.to_s, key, 4096, 8, 1, 32)
         when :argon2
-          validate_key(key, 32)
-
-          # libsodium can't do 32 byte keys
-          # 4th argument in bytes for libsodium,kilobytes for argon2
-          # RbNaCl::PasswordHash.argon2i(value.to_s, key, 3, 2**26, 32)
-
+          # t_cost, m_cost
           [Argon2::Engine.hash_argon2i(value.to_s, key, 2, 12)].pack("H*")
         when :pbkdf2_hmac
-          # TODO enforce key length (or at least a minimum)
-          # validate_key(32)
-
-          digest = OpenSSL::Digest::SHA256.new
-          OpenSSL::PKCS5.pbkdf2_hmac(value.to_s, key, iterations, digest.digest_length, digest)
+          OpenSSL::PKCS5.pbkdf2_hmac(value.to_s, key, options[:iterations], 32, "sha256")
         else
-          raise ArgumentError, "Unknown algorithm"
+          raise BlindIndex::Error, "Unknown algorithm"
         end
 
-      # encode
-      [value].pack("m")
+      encode = options[:encode]
+      if encode
+        if encode.respond_to?(:call)
+          encode.call(value)
+        else
+          [value].pack("m")
+        end
+      else
+        value
+      end
     end
-  end
-
-  def self.validate_key(key, length)
-    raise BlindIndex::Error, "Key must be #{length} bytes" if key.bytesize != length
-    # raise BlindIndex::Error, "Key must use BINARY encoding" if key.encoding != Encoding::BINARY
   end
 end