+++ test/test_helper.rb
@@ -9,9 +9,6 @@ require "argon2"
 
 ActiveRecord::Base.establish_connection adapter: "sqlite3", database: ":memory:"
 
-# TODO remove after next attr_encryptor release
-ActiveSupport::Deprecation.silenced = true
-
 if ENV["VERBOSE"]
   ActiveRecord::Base.logger = ActiveSupport::Logger.new(STDOUT)
 end
@@ -21,13 +18,15 @@ ActiveRecord::Migration.create_table :users do |t|
   t.string :encrypted_email_iv
   t.string :encrypted_email_bidx
   t.string :encrypted_email_ci_bidx
+  t.binary :encrypted_email_binary_bidx
 end
 
 class User < ActiveRecord::Base
-  attr_encrypted :email, key: "0"*32
+  attr_encrypted :email, key: SecureRandom.random_bytes(32)
 
-  blind_index :email, key: "1"*32
-  blind_index :email_ci, algorithm: :argon2, key: -> { "2"*32 }, attribute: :email, expression: ->(v) { v.try(:downcase) }
+  blind_index :email, key: SecureRandom.random_bytes(32)
+  blind_index :email_ci, algorithm: :scrypt, key: SecureRandom.random_bytes(32), attribute: :email, expression: ->(v) { v.try(:downcase) }
+  blind_index :email_binary, algorithm: :argon2, key: SecureRandom.random_bytes(32), attribute: :email, encode: false
 
   validates :email, uniqueness: true
   validates :email_ci, uniqueness: true